Move mockclock to separate file
